+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 `core` module:
 - Add `Pagination` class to work with pagination
 - Add `Repository` interface with common repository methods
+- Add `Timestamp` value class for representing timestamps
 `core-mit` module
 - Add `DetailsConfiguration` for `WrappedException.kt` and `EncodableResult.kt` to customize exception details inclusion
 - Added `randomBase64()` method to generate random Base64 of the specified length